Restructure stap scripts to allow for more util reuse.
This is a bit painful. We want to put reusable utility functions in a tapset
directory. But need to inject the path to the Python binary we're inspecting.
Otherwise symbol resolution won't work. For now, we have the wrapper script
generate an stpm file containing a globally-accessible macro with that path.
Not sure if there's a better way to accomplish this with SystemTap.